fix incorrect transposition calc in simd impl of Jaro

+++ rapidfuzz/details/simd_avx2.hpp
@@ -601,7 +601,7 @@ static inline native_simd<T> min32(const native_simd<T>& a, const native_simd<T>
 static inline native_simd<uint8_t> sllv(const native_simd<uint8_t>& a,
                                         const native_simd<uint8_t>& count_) noexcept
 {
-    __m256i mask_hi = _mm256_set1_epi32(0xFF00FF00);
+    __m256i mask_hi = _mm256_set1_epi32(static_cast<int32_t>(0xFF00FF00));
     __m256i multiplier_lut = _mm256_set_epi8(0, 0, 0, 0, 0, 0, 0, 0, char(128), 64, 32, 16, 8, 4, 2, 1, 0, 0,
                                              0, 0, 0, 0, 0, 0, char(128), 64, 32, 16, 8, 4, 2, 1);
 
@@ -624,7 +624,7 @@ static inline native_simd<uint8_t> sllv(const native_simd<uint8_t>& a,
 static inline native_simd<uint16_t> sllv(const native_simd<uint16_t>& a,
                                          const native_simd<uint16_t>& count) noexcept
 {
-    const __m256i mask = _mm256_set1_epi32(0xFFFF0000);
+    const __m256i mask = _mm256_set1_epi32(static_cast<int32_t>(0xFFFF0000));
     __m256i low_half = _mm256_sllv_epi32(a, _mm256_andnot_si256(mask, count));
     __m256i high_half = _mm256_sllv_epi32(_mm256_and_si256(mask, a), _mm256_srli_epi32(count, 16));
     return _mm256_blend_epi16(low_half, high_half, 0xAA);
